+++ backend/src/controllers/answers-controller.ts
@@ -15,16 +15,8 @@ async function getById(req: Request, res: Response): Promise<void> {
     // Implementar a lógica para obter uma resposta específica
 }
 
-async function getViolenceState(req: Request, res: Response){
-    // Implementar a lógica para obter todas as respostas
-    const state = await answersService.createStateOccur(req.body);
-    
-    return res.status(httpStatus.CREATED).send(json(state));
-}
-
 export const answersController = {
     create,
     getAll,
     getById,
-    getViolenceState
 }
